Add Windows agent lifecycle commands to run with ovs container
* Added postStart command to check for existence of HNS network
and enable OVS extension if it exists.
* Added preStop command to disable OVS extension to ensure smooth
transition to host network when the pod is being deleted.

This change improves network connectivity during pod upgrade and
restart scenarios. It ensures that the pod can maintain network
connectivity by enabling the OVS extension on existing HNS network,
and properly switch to the host network when agent pod is terminated.

Signed-off-by: Shuyang Xin <gavinx@vmware.com>